Another attempt to make 64bit visual studio build happy..

+++ src/core/soloud_fft.cpp
@@ -106,7 +106,7 @@ static void do_fft_8(float *f, const float *x)
             nbr_coef = 1 << pass;
             h_nbr_coef = nbr_coef >> 1;
             d_nbr_coef = nbr_coef << 1;
-            float *cos_ptr = Soloud_fft_trig_8 + (1 << (pass - 1)) - 4;
+			float *cos_ptr = Soloud_fft_trig_8 + (int)(1 << (pass - 1)) - 4;
 
             do
             {
@@ -229,7 +229,7 @@ static void do_fft_10(float *f, const float *x)
             nbr_coef = 1 << pass;
             h_nbr_coef = nbr_coef >> 1;
             d_nbr_coef = nbr_coef << 1;
-            float *cos_ptr = Soloud_fft_trig_10 + (1 << (pass - 1)) - 4;
+			float *cos_ptr = Soloud_fft_trig_10 + (int)(1 << (pass - 1)) - 4;
             do
             {
                 int i;
@@ -301,7 +301,7 @@ static void do_ifft_8(float *f, float *x)
             nbr_coef = 1 << pass;
             h_nbr_coef = nbr_coef >> 1;
             d_nbr_coef = nbr_coef << 1;
-            float *cos_ptr = Soloud_fft_trig_8 + (1 << (pass - 1)) - 4;
+            float *cos_ptr = Soloud_fft_trig_8 + (int)(1 << (pass - 1)) - 4;
             do
             {
                 int i;
